requirements.txt
python-chess==1.999
Flask==3.1.0
waitress==3.0.2